Conflicts:
	CMakeLists.txt
	core/CMakeLists.txt
	core/admin.c
	core/admin.rl
	core/tarantool.c

+++ test/lib/test_suite.py
@@ -10,20 +10,12 @@ import difflib
 import filecmp
 import shlex
 import time
-from tarantool_silverbox_server import TarantoolSilverboxServer
-from tarantool_connection import AdminConnection, DataConnection
+from server import Server
 import tarantool_preprocessor
 import re
 import cStringIO
 import string
 import traceback
-
-class TestRunException(RuntimeError):
-  """A common exception to use across the program."""
-  def __init__(self, message):
-    self.message = message
-  def __str__(self):
-    return self.message
 
 class FilteredStream:
   """Helper class to filter .result file output"""
@@ -54,27 +46,42 @@ class FilteredStream:
     self.stream.close()
 
 
+def check_valgrind_log(path_to_log):
+  """ Check that there were no warnings in the log."""
+  return os.path.getsize(path_to_log) != 0
+
+
+def print_tail_n(filename, num_lines):
+  """Print N last lines of a file."""
+  with open(filename, "r+") as logfile:
+    tail_n = collections.deque(logfile, num_lines)
+    for line in tail_n:
+      sys.stdout.write(line)
+
+
 class Test:
   """An individual test file. A test object can run itself
   and remembers completion state of the run."""
-  def __init__(self, name, suite_ini):
+  def __init__(self, name, args, suite_ini):
     """Initialize test properties: path to test file, path to
     temporary result file, path to the client program, test status."""
     self.name = name
+    self.args = args
+    self.suite_ini = suite_ini
     self.result = name.replace(".test", ".result")
-    self.tmp_result = os.path.join(suite_ini["vardir"],
+    self.tmp_result = os.path.join(self.args.vardir,
                                    os.path.basename(self.result))
     self.reject = name.replace(".test", ".reject")
-    self.suite_ini = suite_ini
     self.is_executed = False
     self.is_executed_ok = None
     self.is_equal_result = None
+    self.is_valgrind_clean = True
 
   def passed(self):
     """Return true if this test was run successfully."""
     return self.is_executed and self.is_executed_ok and self.is_equal_result
 
-  def run(self):
+  def run(self, server):
     """Execute the test assuming it's a python program.
     If the test aborts, print its output to stdout, and raise
     an exception. Else, comprare result and reject files.
@@ -84,18 +91,10 @@ class Test:
 
     diagnostics = "unknown"
     save_stdout = sys.stdout
-    admin = AdminConnection(self.suite_ini["host"],
-                            self.suite_ini["admin_port"])
-    sql = DataConnection(self.suite_ini["host"],
-                         self.suite_ini["port"])
-    server = self.suite_ini["server"]
     try:
-      admin.connect()
-      sql.connect()
       sys.stdout = FilteredStream(self.tmp_result)
-      server = self.suite_ini["server"]
-      vardir = self.suite_ini["vardir"]
-      execfile(self.name, globals(), locals())
+      stdout_fileno = sys.stdout.stream.fileno()
+      execfile(self.name, dict(locals(), **server.__dict__))
       self.is_executed_ok = True
     except Exception as e:
       traceback.print_exc(e)
@@ -104,15 +103,17 @@ class Test:
       if sys.stdout and sys.stdout != save_stdout:
         sys.stdout.close()
       sys.stdout = save_stdout;
-      admin.disconnect()
-      sql.disconnect()
 
     self.is_executed = True
 
     if self.is_executed_ok and os.path.isfile(self.result):
         self.is_equal_result = filecmp.cmp(self.result, self.tmp_result)
 
-    if self.is_executed_ok and self.is_equal_result:
+    if self.args.valgrind:
+      self.is_valgrind_clean = \
+      check_valgrind_log(server.valgrind_log) == False
+
+    if self.is_executed_ok and self.is_equal_result and self.is_valgrind_clean:
       print "[ pass ]"
       os.remove(self.tmp_result)
     elif (self.is_executed_ok and not self.is_equal_result and not
@@ -124,24 +125,28 @@ class Test:
       print "[ fail ]"
       where = ""
       if not self.is_executed_ok:
-        self.print_diagnostics()
+        self.print_diagnostics(self.reject,
+            "Test failed! Last 10 lines of the result file:")
         where = ": test execution aborted, reason '{0}'".format(diagnostics)
-      else:
+      elif not self.is_equal_result:
         self.print_unidiff()
         where = ": wrong test output"
-      if not self.suite_ini["is_force"]:
-        raise TestRunException("Failed to run test " + self.name + where)
+      elif not self.is_valgrind_clean:
+        os.remove(self.reject)
+        self.print_diagnostics(server.valgrind_log,
+                               "Test failed! Last 10 lines of valgrind.log:")
+        where = ": there were warnings in valgrind.log"
+
+      if not self.args.is_force:
+        raise RuntimeError("Failed to run test " + self.name + where)
 
 
-  def print_diagnostics(self):
+  def print_diagnostics(self, logfile, message):
     """Print 10 lines of client program output leading to test
     failure. Used to diagnose a failure of the client program"""
 
-    print "Test failed! Last 10 lines of the result file:"
-    with open(self.reject, "r+") as reject:
-      tail_10 = collections.deque(reject, 10)
-      for line in tail_10:
-        sys.stdout.write(line)
+    print message
+    print_tail_n(logfile, 10)
 
   def print_unidiff(self):
     """Print a unified diff between .test and .result files. Used
@@ -162,20 +167,6 @@ class Test:
         for line in diff:
           sys.stdout.write(line)
 
-class TarantoolConfigFile:
-  """ConfigParser can't read files without sections, work it around"""
-  def __init__(self, fp, section_name):
-    self.fp = fp
-    self.section_name = "[" + section_name + "]"
-  def readline(self):
-    if self.section_name:
-      section_name = self.section_name
-      self.section_name = None
-      return section_name
-    # tarantool.cfg puts string values in quote
-    return self.fp.readline().replace("\"", '')
-
-
 class TestSuite:
   """Each test suite contains a number of related tests files,
   located in the same directory on disk. Each test file has
@@ -195,14 +186,13 @@ class TestSuite:
     self.args = args
     self.tests = []
     self.ini = {}
-    self.ini["suite_path"] = suite_path
-    self.ini["host"] = "localhost"
-    self.ini["is_force"] = self.args.is_force
-    self.ini["vardir"] = args.vardir
+
+    self.ini["core"] = "tarantool"
+    self.ini["module"] = "silverbox"
 
     if os.access(suite_path, os.F_OK) == False:
-      raise TestRunException("Suite \"" + suite_path +\
-                             "\" doesn't exist")
+      raise RuntimeError("Suite \"" + suite_path +\
+                         "\" doesn't exist")
 
 # read the suite config
     config = ConfigParser.ConfigParser()
@@ -213,16 +203,6 @@ class TestSuite:
       self.ini["disabled"] = dict.fromkeys(self.ini["disabled"].split(" "))
     else:
       self.ini["disabled"] = dict()
-# import the necessary module for test suite client
-
-# now read the server config, we need some properties from it
-
-    with open(self.ini["config"]) as fp:
-      dummy_section_name = "tarantool_silverbox"
-      config.readfp(TarantoolConfigFile(fp, dummy_section_name))
-      self.ini["pidfile"] = config.get(dummy_section_name, "pid_file")
-      self.ini["admin_port"] = int(config.get(dummy_section_name, "admin_port"))
-      self.ini["port"] = int(config.get(dummy_section_name, "primary_port"))
 
     print "Collecting tests in \"" + suite_path + "\": " +\
       self.ini["description"] + "."
@@ -230,15 +210,23 @@ class TestSuite:
     for test_name in glob.glob(os.path.join(suite_path, "*.test")):
       for test_pattern in self.args.tests:
         if test_name.find(test_pattern) != -1:
-          self.tests.append(Test(test_name, self.ini))
+          self.tests.append(Test(test_name, self.args, self.ini))
     print "Found " + str(len(self.tests)) + " tests."
 
   def run_all(self):
     """For each file in the test suite, run client program
     assuming each file represents an individual test."""
-    server = TarantoolSilverboxServer(self.args, self.ini)
-    server.install()
-    server.start()
+    try:
+      server = Server(self.ini["core"], self.ini["module"])
+    except Exception as e:
+      print e
+      raise RuntimeError("Unknown server: core = {0}, module = {1}".format(
+        self.ini["core"], self.ini["module"]))
+    server.deploy(self.ini["config"],
+                  server.find_exe(self.args.builddir, silent=False),
+		  self.args.vardir,
+                  self.args.mem, self.args.start_and_exit, self.args.gdb, self.args.valgrind,
+		  silent=False)
     if self.args.start_and_exit:
       print "  Start and exit requested, exiting..."
       exit(0)
@@ -259,7 +247,7 @@ class TestSuite:
       if os.path.basename(test.name) in self.ini["disabled"]:
         print "[ skip ]"
       else:
-        test.run()
+        test.run(server)
         if not test.passed():
           failed_tests.append(test.name)
 
@@ -267,5 +255,11 @@ class TestSuite:
     if len(failed_tests):
       print "Failed {0} tests: {1}.".format(len(failed_tests),
                                             ", ".join(failed_tests))
-    server.stop();
+    server.stop(silent=False)
+    server.cleanup()
 
+    if self.args.valgrind and check_valgrind_log(server.valgrind_log):
+      print "  Error! There were warnings/errors in valgrind log file:"
+      print_tail_n(server.valgrind_log, 20)
+      return 1
+    return len(failed_tests)
